+++ flake.nix
@@ -1,49 +1,65 @@
 {
-  description = "KooL's NixOS-Hyprland"; 
-  	
+  description = "KooL's NixOS-Hyprland";
+
   inputs = {
-	#nixpkgs.url = "github:nixos/nixpkgs/nixos-25.05";
-  	nixpkgs.url = "github:nixos/nixpkgs/nixos-unstable";
-	
-	#hyprland.url = "github:hyprwm/Hyprland"; # hyprland development
-	#distro-grub-themes.url = "github:AdisonCavani/distro-grub-themes";
+    #nixpkgs.url = "github:nixos/nixpkgs/nixos-25.05";
+    nixpkgs.url = "github:nixos/nixpkgs/nixos-unstable";
+
+    #hyprland.url = "github:hyprwm/Hyprland"; # hyprland development
+    #distro-grub-themes.url = "github:AdisonCavani/distro-grub-themes";
+
+    ags = {
+      type = "github";
+      owner = "aylur";
+      repo = "ags";
+      ref = "v1";
+    };
 
     quickshell = {
-        url = "git+https://git.outfoxxed.me/outfoxxed/quickshell";
-            inputs.nixpkgs.follows = "nixpkgs";
-        };
+      url = "git+https://git.outfoxxed.me/outfoxxed/quickshell";
+      inputs.nixpkgs.follows = "nixpkgs";
+    };
 
-  	};
+  };
 
-  outputs = 
-	inputs@{ self, nixpkgs,... }:
-    	let
+  outputs =
+    inputs@{
+      self,
+      nixpkgs,
+      ags,
+      ...
+    }:
+    let
       system = "x86_64-linux";
-      host = "NixOS-Hyprland";
-      username = "alice";
+      host = "default";
+      username = "dwilliams";
 
-    pkgs = import nixpkgs {
-       	inherit system;
-       	config = {
-       	allowUnfree = true;
-       	};
+      pkgs = import nixpkgs {
+        inherit system;
+        config = {
+          allowUnfree = true;
+        };
       };
     in
-      {
-	nixosConfigurations = {
-      "${host}" = nixpkgs.lib.nixosSystem rec {
-		specialArgs = { 
-			inherit system;
-			inherit inputs;
-			inherit username;
-			inherit host;
-			};
-	   		modules = [ 
-				./hosts/${host}/config.nix 
-				# inputs.distro-grub-themes.nixosModules.${system}.default
-                ./modules/quickshell.nix  # quickshell module
-				];
-			};
-		};
-	};
+    {
+      nixosConfigurations = {
+        "${host}" = nixpkgs.lib.nixosSystem rec {
+          specialArgs = {
+            inherit system;
+            inherit inputs;
+            inherit username;
+            inherit host;
+          };
+          modules = [
+            ./hosts/${host}/config.nix
+            # inputs.distro-grub-themes.nixosModules.${system}.default
+            ./modules/quickshell.nix # quickshell module
+            ./modules/packages.nix # Software packages
+            ./modules/fonts.nix # Fonts packages
+            ./modules/portals.nix # portal
+            ./modules/theme.nix # Set dark theme
+          ];
+        };
+      };
+    };
 }